use int64_t for slot in AP_GetPlayerFromSlot

+++ Archipelago.cpp
@@ -762,7 +762,7 @@ const char* AP_GetLocationItemPlayer(AP_State* state, int64_t location_id) {
     return state->location_item_player[location_id].c_str();
 }
 
-const char* AP_GetPlayerFromSlot(AP_State* state, int slot) {
+const char* AP_GetPlayerFromSlot(AP_State* state, int64_t slot) {
     return state->map_players[slot].alias.c_str();
 }
